django/requirements.txt
# django/requirements.txt

# ------------------------------------------------------------------
# Django Core & WSGI Server
# ------------------------------------------------------------------
Django==4.2.1
gunicorn==20.1.0

# ------------------------------------------------------------------
# Database Driver
# ------------------------------------------------------------------
psycopg2-binary==2.9.9

# ------------------------------------------------------------------
# Django REST Framework (DRF) & Auth & CORS
# ------------------------------------------------------------------
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-extensions==3.2.3  # 💡 必須: これが不足していたためエラーが出ていました

# ------------------------------------------------------------------
# Environment Variables & Data Processing
# ------------------------------------------------------------------
python-dotenv==1.0.1
tqdm

# ------------------------------------------------------------------
# API Communication & HTTP Clients
# ------------------------------------------------------------------
requests==2.31.0
urllib3==2.1.0

whitenoise==6.5.0